修复自瞄遇0爆炸问题

+++ User/Middleware/Algorithm/alg_pid.h
@@ -37,23 +37,23 @@ enum Enum_PID_D_First
 class Class_PID
 {
 public:
-    void Init(float __K_P, float __K_I, float __K_D, float __K_F = 0.0f, float __I_Out_Max = 0.0f, float __Out_Max = 0.0f,  float __I_Variable_Speed_A = 0.0f, float __I_Variable_Speed_B = 0.0f, float __I_Separate_Threshold = 0.0f, float __D_T = 0.001f, float __Dead_Zone = 0.0f,Enum_PID_D_First __D_First = PID_D_First_DISABLE);
+    void Init(double __K_P, double __K_I, double __K_D, double __K_F = 0.0f, double __I_Out_Max = 0.0f, double __Out_Max = 0.0f,  double __I_Variable_Speed_A = 0.0f, double __I_Variable_Speed_B = 0.0f, double __I_Separate_Threshold = 0.0f, double __D_T = 0.001f, double __Dead_Zone = 0.0f,Enum_PID_D_First __D_First = PID_D_First_DISABLE);
 
-    inline float Get_Integral_Error();
-    inline float Get_Out();
+    inline double Get_Integral_Error();
+    inline double Get_Out();
 
-    inline void Set_K_P(float __K_P);
-    inline void Set_K_I(float __K_I);
-    inline void Set_K_D(float __K_D);
-    inline void Set_K_F(float __K_F);
-    inline void Set_I_Out_Max(float __I_Out_Max);
-    inline void Set_Out_Max(float __Out_Max);
-    inline void Set_I_Variable_Speed_A(float __Variable_Speed_I_A);
-    inline void Set_I_Variable_Speed_B(float __Variable_Speed_I_B);
-    inline void Set_I_Separate_Threshold(float __I_Separate_Threshold);
-    inline void Set_Target(float __Target);
-    inline void Set_Now(float __Now);
-    inline void Set_Integral_Error(float __Integral_Error);
+    inline void Set_K_P(double __K_P);
+    inline void Set_K_I(double __K_I);
+    inline void Set_K_D(double __K_D);
+    inline void Set_K_F(double __K_F);
+    inline void Set_I_Out_Max(double __I_Out_Max);
+    inline void Set_Out_Max(double __Out_Max);
+    inline void Set_I_Variable_Speed_A(double __Variable_Speed_I_A);
+    inline void Set_I_Variable_Speed_B(double __Variable_Speed_I_B);
+    inline void Set_I_Separate_Threshold(double __I_Separate_Threshold);
+    inline void Set_Target(double __Target);
+    inline void Set_Now(double __Now);
+    inline void Set_Integral_Error(double __Integral_Error);
 
     void TIM_Adjust_PeriodElapsedCallback();
 
@@ -61,9 +61,9 @@ protected:
     //初始化相关常量
 
     // PID计时器周期, s
-    float D_T;
+    double D_T;
     //死区, Error在其绝对值内不输出
-    float Dead_Zone;
+    double Dead_Zone;
     //微分先行
     Enum_PID_D_First D_First;
 
@@ -72,51 +72,51 @@ protected:
     //内部变量
 
     //之前的当前值
-    float Pre_Now = 0.0f;
+    double Pre_Now = 0.0f;
     //之前的目标值
-    float Pre_Target = 0.0f;
+    double Pre_Target = 0.0f;
     //之前的输出值
-    float Pre_Out = 0.0f;
+    double Pre_Out = 0.0f;
     //前向误差
-    float Pre_Error = 0.0f;
+    double Pre_Error = 0.0f;
 
     //读变量
 
     //输出值
-    float Out = 0.0f;
+    double Out = 0.0f;
 
     //写变量
 
     // PID的P
-    float K_P = 0.0f;
+    double K_P = 0.0f;
     // PID的I
-    float K_I = 0.0f;
+    double K_I = 0.0f;
     // PID的D
-    float K_D = 0.0f;
+    double K_D = 0.0f;
     //前馈
-    float K_F = 0.0f;
+    double K_F = 0.0f;
 
     //积分限幅, 0为不限制
-    float I_Out_Max = 0;
+    double I_Out_Max = 0;
     //输出限幅, 0为不限制
-    float Out_Max = 0;
+    double Out_Max = 0;
 
     //变速积分定速内段阈值, 0为不限制
-    float I_Variable_Speed_A = 0.0f;
+    double I_Variable_Speed_A = 0.0f;
     //变速积分变速区间, 0为不限制
-    float I_Variable_Speed_B = 0.0f;
+    double I_Variable_Speed_B = 0.0f;
     //积分分离阈值，需为正数, 0为不限制
-    float I_Separate_Threshold = 0.0f;
+    double I_Separate_Threshold = 0.0f;
 
     //目标值
-    float Target = 0.0f;
+    double Target = 0.0f;
     //当前值
-    float Now = 0.0f;
+    double Now = 0.0f;
 
     //读写变量
 
     //积分值
-    float Integral_Error = 0.0f;
+    double Integral_Error = 0.0f;
 
     //内部函数
 };
@@ -128,9 +128,9 @@ protected:
 /**
  * @brief 获取输出值
  *
- * @return float 输出值
+ * @return double 输出值
  */
-float Class_PID::Get_Integral_Error()
+double Class_PID::Get_Integral_Error()
 {
     return (Integral_Error);
 }
@@ -138,9 +138,9 @@ float Class_PID::Get_Integral_Error()
 /**
  * @brief 获取输出值
  *
- * @return float 输出值
+ * @return double 输出值
  */
-float Class_PID::Get_Out()
+double Class_PID::Get_Out()
 {
     return (Out);
 }
@@ -150,7 +150,7 @@ float Class_PID::Get_Out()
  *
  * @param __K_P PID的P
  */
-void Class_PID::Set_K_P(float __K_P)
+void Class_PID::Set_K_P(double __K_P)
 {
     K_P = __K_P;
 }
@@ -160,7 +160,7 @@ void Class_PID::Set_K_P(float __K_P)
  *
  * @param __K_I PID的I
  */
-void Class_PID::Set_K_I(float __K_I)
+void Class_PID::Set_K_I(double __K_I)
 {
     K_I = __K_I;
 }
@@ -170,7 +170,7 @@ void Class_PID::Set_K_I(float __K_I)
  *
  * @param __K_D PID的D
  */
-void Class_PID::Set_K_D(float __K_D)
+void Class_PID::Set_K_D(double __K_D)
 {
     K_D = __K_D;
 }
@@ -180,7 +180,7 @@ void Class_PID::Set_K_D(float __K_D)
  *
  * @param __K_D 前馈
  */
-void Class_PID::Set_K_F(float __K_F)
+void Class_PID::Set_K_F(double __K_F)
 {
     K_F = __K_F;
 }
@@ -190,7 +190,7 @@ void Class_PID::Set_K_F(float __K_F)
  *
  * @param __I_Out_Max 积分限幅, 0为不限制
  */
-void Class_PID::Set_I_Out_Max(float __I_Out_Max)
+void Class_PID::Set_I_Out_Max(double __I_Out_Max)
 {
     I_Out_Max = __I_Out_Max;
 }
@@ -200,7 +200,7 @@ void Class_PID::Set_I_Out_Max(float __I_Out_Max)
  *
  * @param __Out_Max 输出限幅, 0为不限制
  */
-void Class_PID::Set_Out_Max(float __Out_Max)
+void Class_PID::Set_Out_Max(double __Out_Max)
 {
     Out_Max = __Out_Max;
 }
@@ -210,7 +210,7 @@ void Class_PID::Set_Out_Max(float __Out_Max)
  *
  * @param __I_Variable_Speed_A 定速内段阈值, 0为不限制
  */
-void Class_PID::Set_I_Variable_Speed_A(float __I_Variable_Speed_A)
+void Class_PID::Set_I_Variable_Speed_A(double __I_Variable_Speed_A)
 {
     I_Variable_Speed_A = __I_Variable_Speed_A;
 }
@@ -220,7 +220,7 @@ void Class_PID::Set_I_Variable_Speed_A(float __I_Variable_Speed_A)
  *
  * @param __I_Variable_Speed_B 变速区间, 0为不限制
  */
-void Class_PID::Set_I_Variable_Speed_B(float __I_Variable_Speed_B)
+void Class_PID::Set_I_Variable_Speed_B(double __I_Variable_Speed_B)
 {
     I_Variable_Speed_B = __I_Variable_Speed_B;
 }
@@ -230,7 +230,7 @@ void Class_PID::Set_I_Variable_Speed_B(float __I_Variable_Speed_B)
  *
  * @param __I_Separate_Threshold 积分分离阈值，需为正数, 0为不限制
  */
-void Class_PID::Set_I_Separate_Threshold(float __I_Separate_Threshold)
+void Class_PID::Set_I_Separate_Threshold(double __I_Separate_Threshold)
 {
     I_Separate_Threshold = __I_Separate_Threshold;
 }
@@ -240,7 +240,7 @@ void Class_PID::Set_I_Separate_Threshold(float __I_Separate_Threshold)
  *
  * @param __Target 目标值
  */
-void Class_PID::Set_Target(float __Target)
+void Class_PID::Set_Target(double __Target)
 {
     Target = __Target;
 }
@@ -250,7 +250,7 @@ void Class_PID::Set_Target(float __Target)
  *
  * @param __Now 当前值
  */
-void Class_PID::Set_Now(float __Now)
+void Class_PID::Set_Now(double __Now)
 {
     Now = __Now;
 }
@@ -260,7 +260,7 @@ void Class_PID::Set_Now(float __Now)
  *
  * @param __Set_Integral_Error 积分值
  */
-void Class_PID::Set_Integral_Error(float __Integral_Error)
+void Class_PID::Set_Integral_Error(double __Integral_Error)
 {
     Integral_Error = __Integral_Error;
 }
